requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
google-adk==1.26.0
httpx==0.28.1
livekit-api==1.1.0
livekit-agents==1.4.5
